+++ Lost Items S.A/Assets/Scripts/CommandController.cs
@@ -7,7 +7,7 @@ public class CommandController : MonoBehaviour
 
     FlamaTimer commandTimer;
     List<LostObject.LostObjectType> commandItems = new List<LostObject.LostObjectType>();
-    public float commandScore = 0f;
+    public int commandScore = 0;
     
     public void StartCommand(float time, List<LostObject.LostObjectType> items)
     {
@@ -17,7 +17,7 @@ public class CommandController : MonoBehaviour
         items = commandItems;
         items.Sort();
 
-        commandScore = 300f + items.Count * 100f;
+        commandScore = 300 + items.Count * 100;
     }
 
     public bool UpdateCommand()
